create xib custom cell for collection view
