+++ Travl.iOS/View/DiscoverCell/DiscoverCell.xib
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="collection view cell content view" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="DiscoverCell" customModule="Travl_iOS" customModuleProvider="target"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="discoverCell" id="nPm-kF-92E" customClass="DiscoverCell" customModule="Travl_iOS" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="180" height="215"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="Qgo-tq-ptl">
+                <rect key="frame" x="0.0" y="0.0" width="180" height="215"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="mjH-cA-AAk">
+                        <rect key="frame" x="0.0" y="0.0" width="180" height="184.5"/>
+                        <constraints>
+                            <constraint firstAttribute="width" constant="180" id="B49-TQ-HpU"/>
+                            <constraint firstAttribute="height" constant="184.5" id="mgR-cb-KiL"/>
+                        </constraints>
+                    </imageView>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="3EC-ga-biG">
+                        <rect key="frame" x="5" y="194.5" width="175" height="15.5"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+                <color key="backgroundColor" systemColor="tertiarySystemBackgroundColor"/>
+                <constraints>
+                    <constraint firstItem="mjH-cA-AAk" firstAttribute="leading" secondItem="Qgo-tq-ptl" secondAttribute="leading" id="3q0-oQ-14K"/>
+                    <constraint firstAttribute="trailing" secondItem="3EC-ga-biG" secondAttribute="trailing" id="AYX-Xm-Rvk"/>
+                    <constraint firstAttribute="trailing" secondItem="mjH-cA-AAk" secondAttribute="trailing" id="RnF-Fo-pyV"/>
+                    <constraint firstItem="3EC-ga-biG" firstAttribute="top" secondItem="mjH-cA-AAk" secondAttribute="bottom" constant="10" id="S6S-hV-MEw"/>
+                    <constraint firstItem="3EC-ga-biG" firstAttribute="leading" secondItem="Qgo-tq-ptl" secondAttribute="leading" constant="5" id="dyE-VW-ZYi"/>
+                    <constraint firstItem="mjH-cA-AAk" firstAttribute="top" secondItem="Qgo-tq-ptl" secondAttribute="top" id="sMs-yr-udu"/>
+                    <constraint firstAttribute="bottom" secondItem="3EC-ga-biG" secondAttribute="bottom" constant="5" id="wrp-WY-pq7"/>
+                </constraints>
+            </collectionViewCellContentView>
+            <connections>
+                <outlet property="locationImage" destination="mjH-cA-AAk" id="tBG-7Z-GZi"/>
+                <outlet property="locationLabel" destination="3EC-ga-biG" id="cjq-ak-vHM"/>
+            </connections>
+            <point key="canvasLocation" x="-442.02898550724643" y="176.45089285714286"/>
+        </collectionViewCell>
+    </objects>
+    <resources>
+        <systemColor name="tertiarySystemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
