+++ Travl.iOS/View/Travl.iOS/View/Base.lproj/Discover.storyboard
@@ -30,46 +30,13 @@
                                     <inset key="sectionInset" minX="20" minY="0.0" maxX="20" maxY="0.0"/>
                                 </collectionViewFlowLayout>
                                 <cells>
-                                    <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="cell" id="v5D-aw-2tf" customClass="DiscoverCell" customModule="Travl_iOS" customModuleProvider="target">
-                                        <rect key="frame" x="20" y="0.0" width="178" height="214"/>
+                                    <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="toy-WV-uoE">
+                                        <rect key="frame" x="20" y="0.0" width="128" height="128"/>
                                         <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                                        <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="yt0-UB-TN5">
-                                            <rect key="frame" x="0.0" y="0.0" width="178" height="214"/>
+                                        <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="SyL-l7-hCf">
+                                            <rect key="frame" x="0.0" y="0.0" width="128" height="128"/>
                                             <autoresizingMask key="autoresizingMask"/>
-                                            <subviews>
-                                                <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="qtQ-B3-zal">
-                                                    <rect key="frame" x="0.0" y="0.0" width="178" height="188"/>
-                                                    <constraints>
-                                                        <constraint firstAttribute="width" constant="178" id="LiB-C8-nmW"/>
-                                                        <constraint firstAttribute="height" constant="180" id="Rhh-fO-gm1"/>
-                                                    </constraints>
-                                                </imageView>
-                                                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="edL-4d-Ctn">
-                                                    <rect key="frame" x="0.0" y="193" width="178" height="21"/>
-                                                    <constraints>
-                                                        <constraint firstAttribute="height" constant="21" id="aKY-11-iXE"/>
-                                                        <constraint firstAttribute="width" constant="178" id="odm-Uf-hDi"/>
-                                                    </constraints>
-                                                    <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                                    <nil key="textColor"/>
-                                                    <nil key="highlightedColor"/>
-                                                </label>
-                                            </subviews>
-                                            <constraints>
-                                                <constraint firstAttribute="trailing" secondItem="edL-4d-Ctn" secondAttribute="trailing" id="0ds-7w-eg5"/>
-                                                <constraint firstItem="edL-4d-Ctn" firstAttribute="leading" secondItem="yt0-UB-TN5" secondAttribute="leading" id="ROk-yb-Uiy"/>
-                                                <constraint firstAttribute="bottom" secondItem="edL-4d-Ctn" secondAttribute="bottom" id="UBr-Z5-DDj"/>
-                                                <constraint firstItem="qtQ-B3-zal" firstAttribute="leading" secondItem="yt0-UB-TN5" secondAttribute="leading" id="XBX-ts-zbg"/>
-                                                <constraint firstAttribute="trailing" secondItem="qtQ-B3-zal" secondAttribute="trailing" id="gbC-CT-FpW"/>
-                                                <constraint firstItem="edL-4d-Ctn" firstAttribute="top" secondItem="qtQ-B3-zal" secondAttribute="bottom" constant="5" id="oIK-zP-Gd0"/>
-                                                <constraint firstItem="qtQ-B3-zal" firstAttribute="top" secondItem="yt0-UB-TN5" secondAttribute="top" id="p0R-kA-7kb"/>
-                                            </constraints>
                                         </collectionViewCellContentView>
-                                        <size key="customSize" width="178" height="214"/>
-                                        <connections>
-                                            <outlet property="locationImage" destination="qtQ-B3-zal" id="NFr-JT-sal"/>
-                                            <outlet property="locationLabel" destination="edL-4d-Ctn" id="6mC-6j-6HN"/>
-                                        </connections>
                                     </collectionViewCell>
                                 </cells>
                             </collectionView>
@@ -90,7 +57,7 @@
         <!--  -->
         <scene sceneID="MGL-T5-ej9">
             <objects>
-                <viewController id="Pcy-NL-ews" customClass="DiscoverDetailVC" customModule="Travl_iOS" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="Pcy-NL-ews" customClass="ItenaryVC" customModule="Travl_iOS" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="Zxq-VB-ISR">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
